Break up context
