+++ browser/chrome-extension/src/popup/popup.tsx
@@ -1,14 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { UIProvider } from "../shared/UIContext";
+import ContextProviders from "../shared/contexts/ContextProviders";
 import PopupContent from "./PopupContent";
 import "../shared/extension.css";
 
 const PopupUI = () => {
     return (
-        <UIProvider>
+        <ContextProviders isBrowser={false}>
             <PopupContent />
-        </UIProvider>
+        </ContextProviders>
     );
 };
